Add files via upload
This is the inital commit of my company's website. 

This first commit contains the homepage. It has a static leftsection with the company name and a navbar. The right section contains a scrolling section with projects I have done. Each project contains a stylized picture and a few sentences what the project is about. The website is written in Dutch. 

I've tried to make the website scale with different window sizes by using viewport units and percentages.

This is my first ever project and my first real commit.